Refactoring + fxml for snapshot table
