+++ app/save-and-restore/app/src/main/java/org/phoebus/applications/saveandrestore/ui/snapshot/SnapshotTableViewController.java
@@ -1,0 +1,117 @@
+/*
+ * Copyright (C) 2020 European Spallation Source ERIC.
+ *
+ *  This program is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU General Public License
+ *  as published by the Free Software Foundation; either version 2
+ *  of the License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program; if not, write to the Free Software
+ *  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ *
+ */
+
+package org.phoebus.applications.saveandrestore.ui.snapshot;
+
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.scene.control.CheckBox;
+import javafx.scene.control.ContextMenu;
+import javafx.scene.control.MenuItem;
+import javafx.scene.control.SelectionMode;
+import javafx.scene.control.SeparatorMenuItem;
+import javafx.scene.control.TableRow;
+import javafx.scene.control.TableView;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import org.phoebus.applications.saveandrestore.SaveAndRestoreApplication;
+import org.phoebus.core.types.TimeStampedProcessVariable;
+import org.phoebus.framework.selection.SelectionService;
+import org.phoebus.ui.application.ContextMenuHelper;
+
+import java.time.Instant;
+import java.util.List;
+import java.util.stream.Collectors;
+
+
+public class SnapshotTableViewController {
+
+    @FXML
+    private TableView<TableEntry> snapshotTableView;
+
+    @FXML
+    public void initialize(){
+
+        snapshotTableView.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
+        snapshotTableView.setMaxWidth(Double.MAX_VALUE);
+        snapshotTableView.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
+        snapshotTableView.getStylesheets().add(SnapshotTableViewController.class.getResource("/save-and-restore-style.css").toExternalForm());
+
+        snapshotTableView.addEventHandler(KeyEvent.KEY_PRESSED, event -> {
+            if (event.getCode() != KeyCode.SPACE) {
+                return;
+            }
+
+            ObservableList<TableEntry> selections = snapshotTableView.getSelectionModel().getSelectedItems();
+
+            if (selections == null) {
+                return;
+            }
+
+            selections.stream().filter(item -> !item.readOnlyProperty().get()).forEach(item -> item.selectedProperty().setValue(!item.selectedProperty().get()));
+
+            // Somehow JavaFX TableView handles SPACE pressed event as going into edit mode of the cell.
+            // Consuming event prevents NullPointerException.
+            event.consume();
+        });
+
+        snapshotTableView. setRowFactory(tableView -> new TableRow<>() {
+            final ContextMenu contextMenu = new ContextMenu();
+
+            @Override
+            protected void updateItem(TableEntry item, boolean empty) {
+                super.updateItem(item, empty);
+                if (item == null || empty) {
+                    setOnContextMenuRequested(null);
+                } else {
+                    setOnContextMenuRequested(event -> {
+                        List<TimeStampedProcessVariable> selectedPVList = snapshotTableView.getSelectionModel().getSelectedItems().stream()
+                                .map(tableEntry -> {
+                                    Instant time = Instant.now();
+                                    if (tableEntry.liveTimestampProperty().getValue() != null) {
+                                        time = tableEntry.liveTimestampProperty().getValue();
+                                    }
+                                    return new TimeStampedProcessVariable(tableEntry.pvNameProperty().get(), time);
+                                })
+                                .collect(Collectors.toList());
+
+                        contextMenu.hide();
+                        contextMenu.getItems().clear();
+                        SelectionService.getInstance().setSelection(SaveAndRestoreApplication.NAME, selectedPVList);
+                        ContextMenuHelper.addSupportedEntries(this, contextMenu);
+                        contextMenu.getItems().add(new SeparatorMenuItem());
+                        MenuItem toggle = new MenuItem();
+                        toggle.setText(item.readOnlyProperty().get() ? "Make restorable" : "Make readonly");
+                        CheckBox toggleIcon = new CheckBox();
+                        toggleIcon.setFocusTraversable(false);
+                        toggleIcon.setSelected(item.readOnlyProperty().get());
+                        toggle.setGraphic(toggleIcon);
+                        toggle.setOnAction(actionEvent -> {
+                            item.readOnlyProperty().setValue(!item.readOnlyProperty().get());
+                            item.selectedProperty().set(!item.readOnlyProperty().get());
+                        });
+                        contextMenu.getItems().add(toggle);
+                        contextMenu.show(this, event.getScreenX(), event.getScreenY());
+                    });
+                }
+            }
+        });
+    }
+
+}
