+++ app/save-and-restore/app/src/main/resources/org/phoebus/applications/saveandrestore/ui/snapshot/SnapshotTableView.fxml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+  ~ Copyright (C) 2020 European Spallation Source ERIC.
+  ~
+  ~  This program is free software; you can redistribute it and/or
+  ~  modify it under the terms of the GNU General Public License
+  ~  as published by the Free Software Foundation; either version 2
+  ~  of the License, or (at your option) any later version.
+  ~  
+  ~  This program is distributed in the hope that it will be useful,
+  ~  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~  GNU General Public License for more details.
+  ~  
+  ~  You should have received a copy of the GNU General Public License
+  ~  along with this program; if not, write to the Free Software
+  ~  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+  ~  
+  -->
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import org.phoebus.applications.saveandrestore.ui.snapshot.TooltipTableColumn?>
+
+<AnchorPane maxWidth="1.7976931348623157E308" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.phoebus.applications.saveandrestore.ui.snapshot.SnapshotTableViewController">
+    <TableView fx:id="snapshotTableView" maxWidth="1.7976931348623157E308">
+        <columns>
+            <TableColumn fx:id="selectedColumn" text="" />
+            <TooltipTableColumn fx:id="indexColumn" text="#" />
+            <TableColumn fx:id="pvNameColumn" text="PV Name" />
+            <TableColumn fx:id="radbackPvName" text="Readback PV Name" visible="false" />
+            <TableColumn text="Stored">
+                <columns>
+                    <TableColumn fx:id="storedTimeColumn" text="Time" />
+                    <TableColumn fx:id="storedStatusColumn" text="Status" />
+                    <TableColumn fx:id="storedSeverityColumn" text="Severity" />
+                </columns>
+            </TableColumn>
+            <TableColumn fx:id="deltaColumn" text="Δ Live Setpoint" />
+            <TableColumn text="Live">
+                <columns>
+                    <TableColumn fx:id="liveTimeColumn" text="Time" />
+                    <TableColumn fx:id="liveStatusColumn" text="Status" />
+                    <TableColumn fx:id="liveSeverityColumn" text="Severity" />
+                </columns>
+            </TableColumn>
+            <TableColumn fx:id="liveReadbackColumn" text="Live Readback" visible="false" />
+        </columns>
+    </TableView>
+</AnchorPane>
